rewritten scoreboard and mostly implemented it for the dungeon

+++ src/dungeon/dungeon_state.rs
@@ -3,7 +3,7 @@
 #[derive(Debug)]
 pub enum DungeonState {
     NotReady,
-    Starting { tick_countdown: u64 },
+    Starting { tick_countdown: u16 },
     Started { current_ticks: u64 }, 
     Finished,
 }
